+++ NEWS.md
@@ -2,6 +2,7 @@
 
 (Github-only release date: 2025-XX-XX)
 -   Update column names recognition from GGIR output.
+-   Update IDs definition removing .RData to match current GGIR versions.
 
 
 # stepmetrics 0.1.2