+++ NEWS.md
@@ -1,6 +1,6 @@
 # stepmetrics 0.1.3
 
-(Github-only release date: 2025-XX-XX)
+(Github-only release date: 2025-09-05)
 -   Update IDs definition removing .RData to match current GGIR versions.
 -   Improve peak cadence test.
 -   Improve extraction of wear time when using GGIR-processed data.